updated readme and final touch done about animations
