+++ README.md
@@ -1,6 +1,11 @@
 # flutter_animations
 
 In this project u  will be able to learn simple animation methods using flutter
+In flutter there is mainly two types
+* [Explicit  Animations :](#explicit-animations) Explicit Animation (self designed Animations )
+* [Implicit Animations :](#implicit-animations) Implicit Animation (flutter in build Animation Styles)
+
+## Explicit Animations
 
 ### Example_1_Animations
 transform rotation through through pivot x,y,z axis
@@ -17,3 +22,62 @@ draw 3D box with  stack translate and rotate using vector 3
 
 <img src="ReadmeGif/3d-box.gif" height="400" >
 
+### Example 4 Hero Animation example
+designed a hero animation that wil decide what will happen to hero widget if pressed
+<p align="center">
+<img  src="ReadmeGif/heroAnimation.gif" height="400"  >
+</p>
+
+### Example 7 custom painter
+
+Created a paiter that paints the  polygons that will paint according to no of sides and animates for each sides and rotaion with radius size
+<p align="center">
+<img  src="ReadmeGif/custompainterwithshape.gif" height="400"  >
+</p>
+
+
+### Example 8 Draggable Drawer Animation
+
+Created a gesture detector to drag the drawer and detect if drag less than 50% delta value of screen reverse animation
+<p align="center">
+<img  src="ReadmeGif/customdrawer animation.gif" height="400"  >
+</p>
+
+
+### Exmple 9 prompBoxAnimation
+
+created a prompt animation with cliping and resize icon and container with slide with offset animation and scale with size animation transition 
+
+<p align="center">
+<img  src="ReadmeGif/ineractiveprompt.gif" height="400"  >
+</p>
+
+
+## Implicit Animations
+Flutter In Built Animations with Animated widgeets
+
+### Example 5 Animated Container
+used implicit animated container to zoomin and out
+
+<p align="center">
+<img  src="ReadmeGif/implicitzoominout.gif" height="400"  >
+</p>
+
+### Example 6 TweenAnimationBuilder to create ColorTween 
+created a circle with clipper then used the Animatedtween to  tween the color 
+<p align="center">
+<img  src="ReadmeGif/randomColorTween.gif" height="400"  >
+</p>
+
+
+
+### additional animations
+tried to create a style simalar to image in a bouncing ball
+<p align="center">
+<img  src="ReadmeGif/bounce.gif" height="400"  >
+</p>
+
+tried to create text animation front and back according to string
+<p align="center">
+<img  src="ReadmeGif/text_animation.gif" height="400"  >
+</p>
